makeStringBuilder & errors moved to utils folder
