+++ src/utils/makeStringBuilder.ts
